se creo ruta auth, y el controlador authControllers, se extrajo validacion para el registro de usuario y se coloco en otro modulo
